1) move the maximum number of Rozofsmount transaction from 32 to 64 2) adjust the buffer allocation of the storcli according to the layout

+++ src/rozofsmount/rozofs_write.c
@@ -1933,7 +1933,7 @@ void rozofs_ll_release_nb(fuse_req_t req, fuse_ino_t ino,
         char fidString[64];
 	rozofs_uuid_unparse(f->fid,fidString);
 	severe("rozofs_ll_release_nb no such inode %llu FID %s",
-	        ino, fidString);
+	        (long long unsigned int)ino, fidString);
         errno = ENOENT;
         goto error;
     }
